Optimized some commands and wrote documentation

+++ src/main/java/biz/princeps/landlord/manager/OfferManager.java
@@ -30,8 +30,10 @@ public class OfferManager {
     }
 
     public void removeOffer(String landname) {
-        offers.remove(landname);
-        Bukkit.getScheduler().scheduleAsyncDelayedTask(plugin, () -> db.remove(landname));
+        if (getOffer(landname) != null) {
+            offers.remove(landname);
+            Bukkit.getScheduler().scheduleAsyncDelayedTask(plugin, () -> db.remove(landname));
+        }
     }
 
 }
